rust two_sum_II update

+++ Two_Pointers/two_sum_II/two_sum_II.rs
@@ -3,13 +3,13 @@
 
 fn two_sum(numbers: &Vec<i32>, target: i32) -> Vec<i32> {
 
-    let mut i: u16 = 0;
-    let mut j = (numbers.len() - 1) as u16;
+    let mut i: usize = 0;
+    let mut j = (numbers.len() - 1) as usize;
     let mut summation: i32;
 
     while i < j {
 
-        summation = numbers[i as usize] + numbers[j as usize];
+        summation = numbers[i] + numbers[j];
 
         if summation == target {
             return vec![(i + 1) as i32, (j + 1) as i32];
